게시판, 메인화면 API 수정

+++ src/main/java/com/example/demo/src/board/BoardDao.java
@@ -18,8 +18,6 @@ public class BoardDao {
 
     /**게시글 작성
      *
-     * @param postBoardReq
-     * @return
      */
     public List<PostBoardRes> postBoard(PostBoardReq postBoardReq){
         String createBoardQuery = "insert into Board (user_idx, post_category, post_name, post_content, post_image, post_view, " +
@@ -47,13 +45,12 @@ public class BoardDao {
 
     /**게시글 전체 조회
      *
-     * @return
      */
-    public List<GetBoardRes> getBoard(){
-        String getBoardResQuery = "SELECT post_idx, Board.user_idx, post_category, post_name, post_content, post_image, post_view, post_recommend, post_comment, " +
-                "post_createAt, post_updateAt, post_status, post_anonymity " +
-                "from badjangDB.Board " +
-                "join User on User.user_idx = Board.user_idx " ;
+    public List<GetBoardRes> getBoard(int user_idx){
+        String getBoardResQuery = "SELECT * ,(select exists(select post_idx, user_idx from Bookmark " +
+                "where post_idx = Board.post_idx " +
+                "and user_idx = ?)) " +
+                "as post_bookmark FROM Board " ;
 
         return this.jdbcTemplate.query(getBoardResQuery,
                 (rs, rowNum) -> new GetBoardRes(
@@ -69,21 +66,21 @@ public class BoardDao {
                         rs.getString("post_createAt"),
                         rs.getString("post_updateAt"),
                         rs.getString("post_status"),
-                        rs.getString("post_anonymity")));
+                        rs.getString("post_anonymity"),
+                        rs.getInt("school_name_idx"),
+                        rs.getInt("post_bookmark")),user_idx);
     }
 
     /**게시글 수정
      *
-     * @param patchBoardReq
-     * @return
      */
     public List<GetBoardRes> patchBoard(PatchBoardReq patchBoardReq){
         String createBoardQuery = "UPDATE Board set post_category = ? , post_name = ? , post_content = ? , post_image = ? , " +
-                "post_updateAt = CURRENT_TIMESTAMP where post_idx = ? " ;
+                "post_updateAt = CURRENT_TIMESTAMP where post_idx = ? and user_idx = ?" ;
 
         Object[] createParams = new Object[]{
                 patchBoardReq.getPost_category(), patchBoardReq.getPost_name(), patchBoardReq.getPost_content(),
-                patchBoardReq.getPost_image(), patchBoardReq.getPost_idx()
+                patchBoardReq.getPost_image(), patchBoardReq.getPost_idx(), patchBoardReq.getUser_idx()
         };
 
         this.jdbcTemplate.update(createBoardQuery, createParams);
@@ -92,14 +89,12 @@ public class BoardDao {
 
     /**게시글 삭제
      *
-     * @param deleteBoardReq
-     * @return
      */
     public List<GetBoardRes> deleteBoard(DeleteBoardReq deleteBoardReq){
-        String deleteBoardQuery = "DELETE from Board where post_idx = ?" ;
+        String deleteBoardQuery = "DELETE from Board where post_idx = ? and user_idx = ?" ;
 
         Object[] deleteParam = new Object[]{
-                deleteBoardReq.getPost_idx()
+                deleteBoardReq.getPost_idx(), deleteBoardReq.getUser_idx()
         };
         this.jdbcTemplate.update(deleteBoardQuery, deleteParam);
         return null;
@@ -107,22 +102,42 @@ public class BoardDao {
 
     /**게시글 상세 조회 (조회수 증감)
      *
-     * @param post_idx
-     * @return
      */
-    public List<GetBoardRes> getBoardDetail(int post_idx){
-        String getBoardQuery = "UPDATE Board set post_view = post_view + 1 " +
+    public GetBoardRes updateViewCount(int post_idx){
+        String updateBoardQuery = "UPDATE Board set post_view = post_view + 1 " +
                 "where post_idx = ? " ;
 
         int updateView = post_idx;
-        this.jdbcTemplate.update(getBoardQuery, updateView);
+        this.jdbcTemplate.update(updateBoardQuery, updateView);
         return null;
+    }
+
+    public List<GetBoardRes> getBoardDetail(int user_idx, int post_idx){
+        String getBoardQuery = "SELECT * ,(select exists(select post_idx, user_idx from Bookmark " +
+                "where post_idx = Board.post_idx " +
+                "and user_idx = ?)) " +
+                "as post_bookmark FROM Board where post_idx = ? " ;
+
+        return this.jdbcTemplate.query(getBoardQuery, (rs, rowNum) -> new GetBoardRes(
+                rs.getInt("post_idx"),
+                rs.getInt("user_idx"),
+                rs.getString("post_category"),
+                rs.getString("post_name"),
+                rs.getString("post_content"),
+                rs.getString("post_image"),
+                rs.getInt("post_view"),
+                rs.getInt("post_recommend"),
+                rs.getInt("post_comment"),
+                rs.getString("post_createAt"),
+                rs.getString("post_updateAt"),
+                rs.getString("post_status"),
+                rs.getString("post_anonymity"),
+                rs.getInt("school_name_idx"),
+                rs.getInt("post_bookmark")),user_idx, post_idx);
     }
 
     /**게시글 상세조회(댓글수 증감)
      *
-     * @param post_idx
-     * @return
      */
     public GetBoardRes updateCommentCount(int post_idx){
         String updateCommentQuery = "UPDATE Board set post_comment = (select count(Comment.post_idx) " +
@@ -134,10 +149,17 @@ public class BoardDao {
         return null;
     }
 
-    /**게시글 상세조회(추천 증가)
+    public GetBoardRes updateAt (int post_idx){
+        String updateCommentQuery = "UPDATE Board set post_updateAt = CURRENT_TIMESTAMP " +
+                "where post_idx = ? " ;
+
+        int updateCommentCount = post_idx;
+        this.jdbcTemplate.update(updateCommentQuery, updateCommentCount);
+        return null;
+    }
+
+    /**게시글 상세조회(추천 증감)
      *
-     * @param post_idx
-     * @return
      */
     public GetBoardRes updateRecommendCount(int post_idx){
         String updateCommentQuery = "UPDATE Board set post_recommend = (select count(Board_Recommend.post_idx) " +
@@ -151,8 +173,6 @@ public class BoardDao {
 
     /**게시글 상세조회(추천 등록)
      *
-     * @param postRecommendReq
-     * @return
      */
     public PostRecommendRes postRecommend(PostRecommendReq postRecommendReq){
         String createRecommendQuery = "INSERT INTO Board_Recommend (user_idx, post_idx)" +
@@ -168,7 +188,6 @@ public class BoardDao {
 
     /**추천 취소
      *
-     * @param postRecommendReq
      */
     public void deleteRecommend(PostRecommendReq postRecommendReq) {
         String deleteRecommend = "DELETE FROM Board_Recommend where post_idx = ? and user_idx = ? " ;
@@ -181,8 +200,6 @@ public class BoardDao {
 
     /**추천 수 중복 여부 체크
      *
-     * @param postRecommendReq
-     * @return
      */
     public int checkRecommend (PostRecommendReq postRecommendReq){
         String checkRecommend = "SELECT COUNT(*) FROM Board_Recommend WHERE post_idx = ? AND user_idx = ?" ;
@@ -193,8 +210,6 @@ public class BoardDao {
 
     /**게시글 댓글 조회
      *
-     * @param post_idx
-     * @return
      */
     public List<GetCommentRes> getComment(int post_idx){
         String getCommentQuery = "SELECT comment_idx, User.user_idx, Board.post_idx, comment_content, comment_recommend, " +
@@ -222,8 +237,6 @@ public class BoardDao {
 
     /**게시글 댓글 작성
      *
-     * @param postCommentReq
-     * @return
      */
     public PostCommentRes postComment(PostCommentReq postCommentReq){
         String createCommentQuery = "INSERT INTO Comment (post_idx ,user_idx, comment_content, comment_recommend, comment_anonymity, " +
@@ -250,8 +263,6 @@ public class BoardDao {
 
     /**댓글 수정
      *
-     * @param patchCommentReq
-     * @return
      */
     public GetCommentRes patchComment(PatchCommentReq patchCommentReq){
         String updateQuery = "UPDATE Comment set comment_content = ? , " +
@@ -278,8 +289,6 @@ public class BoardDao {
 
     /**댓글 삭제
      *
-     * @param deleteCommentReq
-     * @return
      */
     public GetCommentRes deleteComment(DeleteCommentReq deleteCommentReq){
         String deleteQuery = "DELETE from Comment where comment_idx = ? ";
@@ -293,8 +302,6 @@ public class BoardDao {
 
     /**댓글 추천
      *
-     * @param postCommentRecommendReq
-     * @return
      */
     public PostCommentRecommendRes postCommentRecommend(PostCommentRecommendReq postCommentRecommendReq) {
         String createRecommendQuery = "INSERT INTO Comment_Recommend (comment_idx, user_idx)" +
@@ -305,20 +312,10 @@ public class BoardDao {
         };
 
         this.jdbcTemplate.update(createRecommendQuery, createParams);
-
-        String lastInsertIdQuery = "select comment_recommend_idx, comment_idx " +
-                "from badjangDB.Comment_Recommend order by comment_recommend_idx desc limit 1 ";
-
-        return this.jdbcTemplate.queryForObject(lastInsertIdQuery,(rs, rowNum) -> new PostCommentRecommendRes(
-                rs.getInt("comment_recommend_idx"),
-                rs.getInt("comment_idx")
-        ));
+        return null;
     }
 
     /**댓글 추천수
-     *
-     * @param comment_idx
-     * @return
      */
     public GetCommentRes updateCommentRecommendCount(int comment_idx) {
         String updateCommentQuery = "UPDATE Comment set comment_updatedAt = CURRENT_TIMESTAMP, " +
@@ -333,7 +330,6 @@ public class BoardDao {
 
     /**댓글 추천 취소
      *
-     * @param postCommentRecommendReq
      */
     public void deleteCommentRecommend(PostCommentRecommendReq postCommentRecommendReq) {
         String deleteRecommend = "DELETE FROM Comment_Recommend where comment_idx = ? and user_idx = ? " ;
@@ -346,8 +342,6 @@ public class BoardDao {
 
     /**댓글 추천 수 중복 여부 체크
      *
-     * @param postCommentRecommendReq
-     * @return
      */
     public int checkCommentRecommend (PostCommentRecommendReq postCommentRecommendReq){
         String checkRecommend = "SELECT COUNT(*) FROM Comment_Recommend WHERE comment_idx = ? AND user_idx = ?" ;
